11409042044更新: 修正從dal(data)至api層因scaffold db context而產生的警告

+++ DatabaseFirstVersion/Data/Repositories/ClubEventRepository.cs
@@ -49,7 +49,7 @@ namespace SportGroups.Data.Repositories
             {
                 return events = await _context.ClubEvents
                     .AsNoTracking()
-                    .Where(e => e.Sport == condition.Sport)
+                    .Where(e => e.Sport == (int)condition.Sport)
                     .ToListAsync();
             }
